+++ CHANGELOG.rst
@@ -2,8 +2,9 @@
 Changelog for pytest-find-dependencies
 ======================================
 
-Unreleased
-----------
+`0.5.1`_ (2022-01-16)
+---------------------
+Slightly improves test output.
 
 Changes
 ~~~~~~~
@@ -107,6 +108,8 @@ New Features
 .. _`0.3.0`: https://pypi.org/project/pytest-find-dependencies/0.3.0/
 .. _`0.4.0`: https://pypi.org/project/pytest-find-dependencies/0.4.0/
 .. _`0.4.1`: https://pypi.org/project/pytest-find-dependencies/0.4.1/
+.. _`0.5.0`: https://pypi.org/project/pytest-find-dependencies/0.5.0/
+.. _`0.5.1`: https://pypi.org/project/pytest-find-dependencies/0.5.1/
 .. _`#1`: https://github.com/mrbean-bremen/pytest-find-dependencies/issues/1
 .. _`#2`: https://github.com/mrbean-bremen/pytest-find-dependencies/issues/2
 .. _`#3`: https://github.com/mrbean-bremen/pytest-find-dependencies/issues/3